Issue #530: some refactor for the query code
